Move theme XAML files into common assembly
This will make it easier to reuse these themes in other projects.
In addition, I added a copy of the default XAML theme for Windows 8+
(obtained from VS2017 in the DesignTools\SystemThemes\Wpf directory)
to ensure that the UI looks as intended even on downlevel operating
systems that use other themes (namely, Windows 7).